The rest of [1147]:
aMysqlSearch is a working implementation of the new aSearchService API. aMysqlSearch adds search with no external dependencies on other libraries. It is based entirely on Doctrine joins and so it works very efficiently with Doctrine queries that also consider other metadata when deciding what items to display. Performance is good for small to medium sized sites. Rebuilding the index can take a long time on large sites - it is within a factor of 2 or so of the performance of Zend Lucene (which was never terrible in that department really)

** You must run apostrophe:migrate after adding this plugin to your project configuration.

** To achieve searchability for a model other than aPage or aMediaItem you must extend your schema correctly and carry out appropriate migrations. See the end of config/doctrine/schema.yml and the migration code in config/apostropheMysqlSearchPluginConfiguration.class.php to learn how to accomplish this.



git-svn-id: http://svn.apostrophenow.org/plugins/apostropheMysqlSearchPlugin/trunk@4072 c89fab08-2f9b-42bf-8176-bd3ad6bee547